+++ TestApplication/Source/main.cpp
@@ -7,7 +7,7 @@ int main(int argc, char* argv[]) {
 
     const Flashlight::WindowProperties windowProperties {800, 600, "Test Application" };
     
-    app.Run(windowProperties, Flashlight::DebugLevel::Debug);
+    app.Run(windowProperties, DebugLevel::Debug);
 
     return 0;
 }
